Add files via upload
This project is about the implementation of Auto Restart Escalator. There are Escalators located at various stations like malls,railway stations,etc. Unlike regular escalators, this technology continuously monitors the balancing circumstances of escalator passengers. Press the START button the main relay get’s on and current flows through it continuously, the coil continues to get the power and hence Escalator remains ON. Similarly when the STOP button is pressed, it breaks the current to the coil and Escalator turns OFF. Raspberry Pi control the relay. Then once the power is interrupted during the regular process, it registers the power off event and then when power resumes, it registers the power on event. Based on this power on and then back on after sometime, Raspberry is able to start a delay loop for 1 minute, it turns ON the Buzzer and after 10 seconds, then the Relay gets ON. The Relay now supplies the current needed for the coil of Main Power Relay. Thus, the Main Power Relay complete the Lamp circuit and Lamp is turned ON. Same logic also applied for Emergency STOP button.